Remove navigation to "/home" end-point
Remove the call to function navigate('/home'), and add a call to '/' only to
navigate to the root page of the application.

This change is necessary as we haven't decided for a home page yet, only a
book page or a root page can be used interchangeably and here we are just
using "/" for the root end-point.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>