Add component `<LanguageSelector />` to allow users to toggle between languages
This commit includes a "<LanguageSelector />" component to the navigation bar
to allow users to toggle between supported language translations for the content
of "Gita". Currently "<LanguageSelector />" doesn't perform any translation
action because only the design was implemented but in the future commits we'll
implement the routines that will translate the text.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>