Add a close button to the `react-markdown-editor-lite` to fire a close event when user clicks on it
This close event is going to be quite complicated then it looks to be because
as we are using a counter state for creating our editors we are bound to use
the counter state to remove our editors and this can only be done by
decrementing the counter state. But on decrementing the counter state will
pop the editor from the end of the screen not from the place where the close
event had occured.

A possible solution is to bind a id with the text editor and remove the editor
with a specific ID.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>