New component "<MarkdownEditor />" to write markdown and render HTML version of it
The component "<MarkdownEditor />" is especially built for application "Gita"
to provide flexibility to the users to enter markdown and get its HTML version.

A point to be noted is that I've re-written the entire
"react-markdown-editor-lite" library just because I needed some additional
features that were not present in the library by default as a result now we
have a native component to edit and render HTML documents.

Unlike "react-markdown-editor-lite" library our package totally developed using
the functional components and it also support "Plugins" to be added or removed.

There's no change in the preview component we also set the
"dangerouslySetInnerHTML" attribute with the "__html" property to render the
HTML from markdown.

Unfortunately, I did not have time to write tests for this huge package but I've
definitely planned to test this component in future.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>