`text` should change both on changing the `props.text` and `text` itself by the `onChange` event
This commit fixes the problem of the editor strictness over the given "text"
as component property even after triggering the "onChange" event on it manually
which caused the editor component hung-up in one state and never changing to
another.

There's one more fix regarding how the paste event will be handled, the
following conditions:
        configs.allowPasteImage &&
        typeof onImageUpload === 'function'
should only be checked inside of the "uploadWithDataTransfer()" function only
after the "item.kind" given is a "file" and is an "image", otherwise it will
pre-empt pasting string data.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>