High level diagram view of `<Book />` component
Add high level diagram view of the "<Book />" component to make it clear where
to start and where to end implementing the component and connect the components
through the same hooks passed so, when any hook in any of the child component
update itself it will update the parent component too (this time it is <Book />).

Moreover, re-create "<ChapterIndex />" component and separate its design in a
separate "draw.io" sheet in the same file.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>