Add `Firestore` database to `Gita` to achieve seamless data storing facility
A "Add Chapter Button" is necessary in the "<ChapterIndexActionContainer />" to
set new chapters and their relative data directly from the interface without
needing to manually update the firestore database from the Firebase console.

This new "<ChapterIndex />" is completely synced with the database for every add
chapter action triggered from the user end by the means of "firestore-request.js"
module, which is a dedicated module for managing chapters and their contents all
asynchronously.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>