Make `<ChapterIndex />` scrollable to allow scrolling to the bottom list items
In this update we also attach two new attributes to the "chapters" i.e., "position",
and "isSelected". The "position" argument makes sure that every list item is in
order and the "isSelected" attrbute makes sure that on every reload of the page
the reader doesn't lose the last selected page he was in.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>