FIX Change of position of the slide in|out icon with scroll of the `<Book />` component
This commit fixes issue that was missed in the last commit 856f071915555f2e6b100fb322ea636e10fc53a2
which was causing the change in position of the slide in|out icon which was
supposed to be fixed.

Apply "position: fixed;" property to the slide in|out icon and update the css
properties on view change.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>